Refactoring silx-view tests into a dedicated package
