+++ silx/app/view/setup.py
@@ -1,7 +1,6 @@
 # coding: utf-8
 # /*##########################################################################
-#
-# Copyright (c) 2016-2017 European Synchrotron Radiation Facility
+# Copyright (C) 2016 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -21,19 +20,21 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 #
-# ###########################################################################*/
+# ############################################################################*/
+
 __authors__ = ["V. Valls"]
 __license__ = "MIT"
 __date__ = "06/06/2018"
 
-import unittest
-
-from ..view import test as test_view
-from . import test_convert
+from numpy.distutils.misc_util import Configuration
 
 
-def suite():
-    test_suite = unittest.TestSuite()
-    test_suite.addTest(test_view.suite())
-    test_suite.addTest(test_convert.suite())
-    return test_suite
+def configuration(parent_package='', top_path=None):
+    config = Configuration('view', parent_package, top_path)
+    config.add_subpackage('test')
+    return config
+
+
+if __name__ == "__main__":
+    from numpy.distutils.core import setup
+    setup(configuration=configuration)
